requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
future==0.18.2
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
PyJWT==2.0.0
pyparsing==2.4.7
pytz==2020.4
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
